Make vm calls 64bit friendly

+++ trunk/code/cgame/cg_event.c
@@ -606,7 +606,7 @@ typedef struct {
 	int anim;
 } painAnimForTag_t;
 
-#define PEFOFS( x ) ( (int)&( ( (playerEntity_t *)0 )->x ) )
+#define PEFOFS( x ) ( (intptr_t)&( ( (playerEntity_t *)0 )->x ) )
 
 void CG_PainEvent( centity_t *cent, int health, qboolean crouching ) {
 	char    *snd;
