+++ trunk/code/game/bg_animation.c
@@ -1866,7 +1866,7 @@ int BG_GetConditionValue( int client, int condition, qboolean checkConversion ) 
 	int value, i;
 
 	// TTimo gcc: assignment makes integer from pointer without a cast
-	value = (int)globalScriptData->clientConditions[client][condition];
+	value = (intptr_t)globalScriptData->clientConditions[client][condition];
 
 	if ( checkConversion ) {
 		// we may need to convert to a value
